feat: wrapped up workspace caching

+++ apps/api/src/workspace-membership/workspace-membership.service.ts
@@ -12,7 +12,6 @@ import {
   Logger,
   NotFoundException
 } from '@nestjs/common'
-import { JwtService } from '@nestjs/jwt'
 import {
   Authority,
   AuthProvider,
@@ -42,7 +41,6 @@ export class WorkspaceMembershipService {
   constructor(
     private readonly prisma: PrismaService,
     private readonly authorizationService: AuthorizationService,
-    private readonly jwt: JwtService,
     private readonly tierLimitService: TierLimitService,
     @Inject(MAIL_SERVICE) private readonly mailService: IMailService,
     private readonly slugGenerator: SlugGenerator,
@@ -208,7 +206,6 @@ export class WorkspaceMembershipService {
       this.prisma
     )
 
-    await this.workspaceCacheService.removeWorkspaceCache(workspace)
     await this.workspaceCacheService.setWorkspaceAdmin(
       workspace.id,
       otherUser.id
@@ -247,7 +244,7 @@ export class WorkspaceMembershipService {
       throw new BadRequestException(
         constructErrorBody(
           'This workspace has been disabled',
-          'To use the workspace again, remove the previum resources, or upgrade to a paid plan'
+          'To use the workspace again, remove the premium resources, or upgrade to a paid plan'
         )
       )
     }
@@ -303,8 +300,6 @@ export class WorkspaceMembershipService {
       )
       return []
     }
-
-    await this.workspaceCacheService.removeWorkspaceCache(workspace)
   }
 
   /**
@@ -329,59 +324,52 @@ export class WorkspaceMembershipService {
         authorities: [Authority.REMOVE_USER]
       })
 
-    const userIds = await this.prisma.user
-      .findMany({
-        where: {
-          email: {
-            in: userEmails.map((email) => email.toLowerCase())
-          }
-        },
-        select: {
-          id: true
+    const usersToRemove = await this.prisma.user.findMany({
+      where: {
+        email: {
+          in: userEmails.map((email) => email.toLowerCase())
         }
-      })
-      .then((users) => users.map((u) => u.id))
+      },
+      select: {
+        id: true
+      }
+    })
 
+    const userIds = []
     // Remove users from the workspace if any
-    if (userIds && userIds.length > 0) {
-      if (userIds.find((id) => id === user.id)) {
+    for (const userToRemove of usersToRemove) {
+      const userId = userToRemove.id
+      if (userId === user.id) {
         throw new BadRequestException(
           constructErrorBody(
-            `You can not remove yourself from the workspace.`,
-            `You can only leave a workspace.`
+            'Cannot remove yourself from the workspace',
+            'You cannot remove yourself from the workspace'
           )
         )
       }
 
-      // Delete the membership
-      await this.prisma.workspaceMember.deleteMany({
-        where: {
-          workspaceId: workspace.id,
-          userId: {
-            in: userIds
-          }
-        }
-      })
-
-      if (user.emailPreference && !user.emailPreference.activity) {
-        this.log.log(
-          `User ${user.id} has opted out of receiving activity notifications`
-        )
-        return
-      }
-
-      // Send an email to the removed users
-      const removedOn = new Date()
-      const emailPromises = userEmails.map((userEmail) =>
-        this.mailService.removedFromWorkspace(
-          userEmail,
-          workspace.name,
-          removedOn
-        )
-      )
-
-      await Promise.all(emailPromises)
+      userIds.push(userId)
+      await this.deleteMembership(workspace, userId)
     }
+
+    if (user.emailPreference && !user.emailPreference.activity) {
+      this.log.log(
+        `User ${user.id} has opted out of receiving activity notifications`
+      )
+      return
+    }
+
+    // Send email to the removed users
+    const removedOn = new Date()
+    const emailPromises = userEmails.map((userEmail) =>
+      this.mailService.removedFromWorkspace(
+        userEmail,
+        workspace.name,
+        removedOn
+      )
+    )
+
+    await Promise.all(emailPromises)
 
     await createEvent(
       {
@@ -403,8 +391,6 @@ export class WorkspaceMembershipService {
     this.log.debug(
       `Removed users from workspace ${workspace.name} (${workspace.id})`
     )
-
-    await this.workspaceCacheService.removeWorkspaceCache(workspace)
   }
 
   /**
@@ -678,7 +664,7 @@ export class WorkspaceMembershipService {
       throw new BadRequestException(
         constructErrorBody(
           'This workspace has been disabled',
-          'To use the workspace again, remove the previum resources, or upgrade to a paid plan'
+          'To use the workspace again, remove the premium resources, or upgrade to a paid plan'
         )
       )
     }
@@ -749,7 +735,7 @@ export class WorkspaceMembershipService {
     await this.checkInvitationPending(workspaceSlug, inviteeUser)
 
     // Delete the membership
-    await this.deleteMembership(workspace.id, inviteeUser.id)
+    await this.deleteMembership(workspace, inviteeUser.id)
 
     await createEvent(
       {
@@ -770,8 +756,6 @@ export class WorkspaceMembershipService {
     this.log.debug(
       `User ${user.name} (${user.id}) cancelled invitation to workspace ${workspace.id}`
     )
-
-    await this.workspaceCacheService.removeWorkspaceCache(workspace)
   }
 
   /**
@@ -796,7 +780,7 @@ export class WorkspaceMembershipService {
     })
 
     // Delete the membership
-    await this.deleteMembership(workspace.id, user.id)
+    await this.deleteMembership(workspace, user.id)
 
     await createEvent(
       {
@@ -816,8 +800,6 @@ export class WorkspaceMembershipService {
     this.log.debug(
       `User ${user.name} (${user.id}) declined invitation to workspace ${workspace.id}`
     )
-
-    await this.workspaceCacheService.removeWorkspaceCache(workspace)
   }
 
   /**
@@ -858,7 +840,7 @@ export class WorkspaceMembershipService {
       )
 
     // Delete the membership
-    await this.deleteMembership(workspace.id, user.id)
+    await this.deleteMembership(workspace, user.id)
 
     await createEvent(
       {
@@ -878,8 +860,6 @@ export class WorkspaceMembershipService {
     this.log.debug(
       `User ${user.name} (${user.id}) left workspace ${workspace.id}`
     )
-
-    await this.workspaceCacheService.removeWorkspaceCache(workspace)
   }
 
   /**
@@ -1059,14 +1039,14 @@ export class WorkspaceMembershipService {
       ) {
         this.log.warn(
           `User ${
-            memberUser.name || memberUser.email
+            memberUser!.name || memberUser!.email
           } (${userId}) is already a member of workspace ${workspace.name} (${
             workspace.slug
           }). Skipping.`
         )
         throw new ConflictException(
           constructErrorBody(
-            `User ${memberUser.name || memberUser.email} is already a member of this workspace`,
+            `User ${memberUser!.name || memberUser!.email} is already a member of this workspace`,
             'Please check the teams tab to confirm whether the user is a member of this workspace'
           )
         )
@@ -1153,6 +1133,11 @@ export class WorkspaceMembershipService {
         })
       )
 
+      await this.workspaceCacheService.addMemberToRawWorkspace(
+        workspace,
+        rawWorkspaceMember.id
+      )
+
       this.log.debug(`Added user ${memberUser} to workspace ${workspace.name}.`)
     }
 
@@ -1203,23 +1188,44 @@ export class WorkspaceMembershipService {
 
   /**
    * Deletes the membership of a user in a workspace.
-   * @param workspaceId The ID of the workspace to delete the membership from
+   * @param workspace
    * @param userId The ID of the user to delete the membership for
    * @returns A promise that resolves when the membership is deleted
    * @private
    */
   private async deleteMembership(
-    workspaceId: Workspace['id'],
+    workspace: Workspace,
     userId: User['id']
   ): Promise<void> {
-    await this.prisma.workspaceMember.delete({
+    const membership = await this.prisma.workspaceMember.findUnique({
       where: {
         workspaceId_userId: {
-          workspaceId,
+          workspaceId: workspace.id,
           userId
         }
       }
     })
+
+    if (!membership) {
+      this.log.error(`Workspace membership not found for user ${userId}`)
+      throw new NotFoundException(
+        constructErrorBody(
+          'Workspace membership not found',
+          'The user is not a member of the workspace'
+        )
+      )
+    }
+
+    await this.prisma.workspaceMember.delete({
+      where: {
+        id: membership.id
+      }
+    })
+
+    await this.workspaceCacheService.removeMemberFromRawWorkspace(
+      workspace,
+      membership.id
+    )
   }
 
   /**
